requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
python-jose
bcrypt==4.0.1
passlib[bcrypt]
redis
colorama